issues/118 perf(UniduxPanelStateTab): optimize list dictionary render
remind type and donot search type repeatedly
